refactor: Пакет client разбит на составные части
