--- app/deps/client.py
@@ -1,15 +1,0 @@
-from services.client import MatrixClient
-from services.client.stream import StreamMatrixClient
-from services.receive.rgb import RGBMessageReceiver
-from services.stream.ws import WsStream
-from starlette.websockets import WebSocket
-
-
-def get_matrix_client(websocket: WebSocket) -> MatrixClient:
-    """Получить объект клиента матрицы для установки и работы с соединением"""
-
-    stream = WsStream(ws=websocket)
-    return StreamMatrixClient(
-        stream,
-        receiver=RGBMessageReceiver(stream),
-    )
